Make table load data based on the selected donation type/status
